requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
colorama==0.4.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pycodestyle==2.7.0
PyJWT==1.7.1
pylint==2.8.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
wrapt==1.12.1
